requirements.txt
Appium-Python-Client~=4.0.0
selenium~=4.21.0
certifi